+++ ChangeLog.md
@@ -1,3 +1,20 @@
+##v 2.0
+- Integrating 3D ribosome structure analysis to ARF
+- All relevant functions now require the rRNA fasta file that is used for mapping.
+- First release of dricARF Differential Ribosome Collision Prediction pipeline
+- Bug fixes (closing issues until now)
+
+##v 1.0
+- Publication release
+- A few bug fixes
+
+##v 0.9.1 - 2022-04-14
+- Paper revision release
+- rRNA-RP contact point distance threshold changed
+- New experimental function that allows altering the Contact Point threshold and possible directionality in predictions
+- A few technical updates
+- RP-rRNA distance matrices are added to the repo
+
 ##v 0.9 - 2021-11-26
 - First release
 - Paper submission version